+++ abacatepay/models/product.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel
+
+class Product(BaseModel):
+    externalId: str
+    name: str
+    quantity: int
+    price: int
+    description: str
